Done some bug fixes and implemented exception handling

+++ quora-service/src/main/java/com/upgrad/quora/service/business/AnswerService.java
@@ -1,7 +1,10 @@
 package com.upgrad.quora.service.business;
 
 import com.upgrad.quora.service.dao.AnswerDao;
+import com.upgrad.quora.service.dao.QuestionDao;
 import com.upgrad.quora.service.entity.AnswerEntity;
+import com.upgrad.quora.service.entity.QuestionEntity;
+import com.upgrad.quora.service.entity.UserAuthTokenEntity;
 import com.upgrad.quora.service.entity.UserEntity;
 import com.upgrad.quora.service.exception.AnswerNotFoundException;
 import com.upgrad.quora.service.exception.AuthorizationFailedException;
@@ -19,6 +22,9 @@ public class AnswerService {
     @Autowired
     AnswerDao answerDao;
 
+    @Autowired
+    QuestionDao questionDao;
+
     @Transactional(propagation = Propagation.REQUIRED)
     public AnswerEntity createAnswer(AnswerEntity answerEntity)  {
         return answerDao.addAnswer(answerEntity);
@@ -26,14 +32,14 @@ public class AnswerService {
 
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public AnswerEntity editAnswer(UserEntity userEntity, String uuid, String editedAnswer) throws AuthorizationFailedException, InvalidQuestionException {
+    public AnswerEntity editAnswer(UserEntity userEntity, String uuid, String editedAnswer) throws AuthorizationFailedException, AnswerNotFoundException {
 
         AnswerEntity answerEntity = answerDao.getAnswerByQuestionId(uuid);
         if(answerEntity==null)
-            throw new InvalidQuestionException("QUES-001", "Entered question uuid does not exist");
+            throw new AnswerNotFoundException("ANS-001", "Entered answer uuid does not exist");
 
         if(!answerEntity.getUuid().equalsIgnoreCase(userEntity.getUuid()))
-            throw new AuthorizationFailedException("ATHR-003", "Only the question owner can edit the question");
+            throw new AuthorizationFailedException("ATHR-003", "Only the answer owner can edit the answer");
 
         answerEntity.setAnswer(editedAnswer);
         answerDao.editAnswer(answerEntity);
@@ -42,11 +48,14 @@ public class AnswerService {
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public AnswerEntity deleteAnswer(String uuid) throws AnswerNotFoundException {
+    public AnswerEntity deleteAnswer(String uuid, UserEntity userEntity, UserAuthTokenEntity userAuthTokenEntity) throws AnswerNotFoundException, AuthorizationFailedException {
 
         AnswerEntity answerEntity = answerDao.getAnswerByQuestionId(uuid);
         if(answerEntity == null)
             throw new AnswerNotFoundException("QUES-001","Entered question uuid does not exist");
+
+        if(answerEntity.getUuid() != userEntity.getUuid() || userAuthTokenEntity.getUuid()!=userEntity.getUuid())
+            throw new AuthorizationFailedException("ATHR-003","Only the question owner can edit the question");
 
         answerDao.deleteQuestion(answerEntity);
 
@@ -54,7 +63,12 @@ public class AnswerService {
     }
 
 
-    public List<AnswerEntity> getAllAnswers(String uuid){
+    public List<AnswerEntity> getAllAnswers(String uuid) throws InvalidQuestionException {
+        QuestionEntity questionEntity = questionDao.getQuestionById(uuid);
+        if(questionEntity==null)
+            throw new InvalidQuestionException("QUES-001","Entered question uuid does not exist");
+
+
         return answerDao.getAllAnswersByQuestionId(uuid);
     }
 }
